Improved workflow

+++ tiff_to_csv.py
@@ -71,11 +71,11 @@ def convert_parquet_to_csv(path_parquet, args):
 
         # get single ID instead of X and Y
         # df.apply(lambda row: row.X + (row.Y - y1) * dx - x1, axis=1)  # size of raster is 86401 x 33601
-        df['ID'] = df["X"] - x1 + (df[ "Y"] - y1)
+        df['ID'] = df["X"] - x1 + (df["Y"] - y1)
         df = df.drop(columns=['X', 'Y'])  # int32 max 2,147,483,647
         # round float values to get lesser size
         # in rade9d step of value is 0.1, therefore 1 decimal is enough
-        logger.debug(f"Before decimal rounding have:\n {df.head(10)}")
+        # logger.debug(f"Before decimal rounding have:\n {df.head(10)}")
         if args.type in ("rade9d",):
             df["VALUE"] = df["VALUE"].round(1)
             logger.debug(f"Round to 1 decimal")
@@ -96,12 +96,18 @@ def convert_parquet_to_csv(path_parquet, args):
         df.to_csv(path_csv, index=False)
         logger.debug(f"Dumped parquet to {path_csv}")
         path_device_csv = os.path.join(folder_device_processed_csv, args.type, filename + ".csv")
+        logger.debug(f"Started uploading of {path_csv}")
         shutil.copyfile(path_csv, path_device_csv)
+        logger.debug(f"Uploaded  {path_csv} to {path_device_csv}")
         if args.delete:
-            os.remove(path_csv)
-            os.remove(path_parquet)
+            if (sys.getsizeof(path_csv) == sys.getsizeof(path_device_csv)):
+                if os.path.exists(path_csv):
+                    os.remove(path_csv)
+                    logger.debug(f"Deleted {path_csv}")
+                if os.path.exists(path_parquet):
+                    os.remove(path_parquet)
+                    logger.debug(f"Deleted {path_parquet}")
         # os.remove(path_parquet)
-        logger.debug(f"Done uploading of {path_csv} to {path_device_csv}")
         return get_first_number_from_string(filename)
     except Exception as exc:
         with open(local_log_file, 'a+') as f:
@@ -110,14 +116,12 @@ def convert_parquet_to_csv(path_parquet, args):
         traceback.print_exc()
         if os.path.exists(path_csv):
             os.remove(path_csv)
+            logger.debug(f"Deleted {path_csv}")
         return None
 
 
 def convert_geotiff_to_parquet(file_path, args):
     """ file_path - path to the .tif file on remote device """
-    #if args.type.startswith("vcm"):
-    #    filename = Path(file_path).name  # with extension
-    #else:
     filename = Path(file_path).stem
     date = get_first_number_from_string(filename)
     if not (args.date_min <= date <= args.date_max):
@@ -129,7 +133,8 @@ def convert_geotiff_to_parquet(file_path, args):
     path_csv = os.path.join(folder_result, filename + ".csv")
     path_parquet = os.path.join(folder_result, filename + ".parquet")
     logger.debug(
-        f"Processing {filename} save results to {os.path.join(folder_result, filename) + '.parquet and/or .csv'}")
+        f"Processing {filename} from {folder_to_process}\n "
+        f"save results to {os.path.join(folder_result, filename) + '.parquet and/or .csv'}")
 
     try:
         # download raster to process
@@ -141,7 +146,7 @@ def convert_geotiff_to_parquet(file_path, args):
             else:
                 logger.debug(f"File {path_raster} is of the right size, no need to download again")
 
-        # get one area - approx Western Europe
+        # get one area - approx Western Europe. Type e.g. XYZ or GTIFF
         com_string = "gdal_translate -of XYZ -q -srcwin " + str(x1) + ", " + str(y1) + ", " + str(
             dx) + ", " + str(dy) + " " + str(path_raster) + " " + str(path_raster_cropped)
         subprocess.run(com_string)
@@ -168,17 +173,22 @@ def convert_geotiff_to_parquet(file_path, args):
         # delete used files
         if os.path.exists(path_raster_cropped):
             os.remove(path_raster_cropped)
+            logger.debug(f"Deleted {path_raster_cropped}")
         if args.delete:
             if os.path.exists(path_raster):
                 os.remove(path_raster)
+                logger.debug(f"Deleted {path_raster}")
         # upload result to device
         path_result_device = os.path.join(folder_device_processed_parquet, args.type, filename + ".parquet")
+        logger.debug(f"Started uploading of {path_parquet}")
         shutil.copyfile(path_parquet, path_result_device)
         logger.debug(f"Uploaded {path_parquet} to {path_result_device}")
         # TODO update list of files - account for python newly updated ones
     except (KeyboardInterrupt, SystemExit):
         if os.path.exists(path_parquet):
-            os.remove(path_parquet)
+            pass
+            # os.remove(path_parquet)
+            # logger.debug(f"Deleted {path_parquet}")
     except Exception as exc:
         with open(local_log_file, 'a+') as f:
             f.write(f"{date} {exc}\n")
@@ -186,6 +196,7 @@ def convert_geotiff_to_parquet(file_path, args):
         traceback.print_exc()
         if os.path.exists(path_parquet):
             os.remove(path_parquet)
+            logger.debug(f"Deleted {path_parquet}")
         return None
 
     try:
@@ -193,6 +204,7 @@ def convert_geotiff_to_parquet(file_path, args):
             convert_parquet_to_csv(path_parquet, args)
             if args.delete and os.path.exists(path_parquet):
                 os.remove(path_parquet)
+                logger.debug(f"Deleted {path_parquet}")
     except Exception as exc:
         with open(local_log_file, 'a+') as f:
             f.write(f"{date} {exc}\n")
@@ -200,6 +212,7 @@ def convert_geotiff_to_parquet(file_path, args):
         traceback.print_exc()
         if os.path.exists(path_csv):
             os.remove(path_csv)
+            logger.debug(f"Deleted {path_csv}")
         return None
     return date
 
@@ -213,48 +226,7 @@ def main(args):
     while True:
         print(time.strftime(time_format))
         dates = []
-        # set general type, to find folder with raw .tif data on the remote device
-        if args.type in ("cloud_cover", "rade9d"):
-            t = "nightly"
-        else:
-            t = "monthly"
 
-        # get filepaths to process
-        if args.extension == ".parquet":
-            if args.from_tgz:
-                pass
-            made_filenames = [Path(f).stem
-                              for f in list(glob.glob(os.path.join(folder_result, "*" + ".parquet")))] \
-                             + [Path(f).stem
-                                for f in
-                                list(glob.glob(os.path.join(folder_device_processed_parquet_current, "*"
-                                                            + ".parquet")))]
-            filepaths = []
-            for filepath_curr in sorted(glob.glob(os.path.join(folder_device_raw, t, args.type, "*.tif"))):
-                filename_curr = Path(filepath_curr).stem
-                if filename_curr not in made_filenames:
-                    filepaths.append(filepath_curr)
-
-            #filepaths = filepaths[0:1]  # TODO delete not in debug
-
-            func = convert_geotiff_to_parquet
-        elif args.extension == ".csv":
-            made_filenames = []  # TODO delete in not debug
-            if 1:
-                made_filenames = [Path(f).stem
-                                  for f in
-                                  list(glob.glob(os.path.join(folder_device_processed_csv, args.type, "*.csv")))]  # \
-            # + [Path(f).stem
-            #   for f in
-            #   list(glob.glob(os.path.join(folder_local_data, args.type, "result", "*" + ".csv")))]
-
-            filepaths = []
-            for filepath_curr in sorted(glob.glob(os.path.join(folder_result, "*.parquet"))):
-                filename_curr = Path(filepath_curr).stem
-                if filename_curr not in made_filenames:
-                    filepaths.append(filepath_curr)
-            func = convert_parquet_to_csv
-        # dates = ([re.findall(r'\d+', Path(f).stem)[0] for f in files])
         if args.parall_type == "process":
             parallel_executor = concurrent.futures.ProcessPoolExecutor(max_workers=args.workers)
         elif args.parall_type == "thread":
@@ -262,6 +234,57 @@ def main(args):
         else:
             raise NotImplementedError
 
+        made_parquets = [Path(f).stem
+                         for f in list(glob.glob(os.path.join(folder_result, "*" + ".parquet")))] \
+                        + [Path(f).stem
+                           for f in
+                           list(glob.glob(os.path.join(folder_device_processed_parquet_current, "*.parquet")))]
+        made_csvs = [Path(f).stem
+                     for f in
+                     list(glob.glob(os.path.join(folder_device_processed_csv_current, "*.csv")))]
+
+        filepaths = []
+        filepaths_tif = []
+        filepaths_parquet = []
+        # TODO better system
+        # .parquet to make from tif parquets, then flag --parquet_to_csv does also to csv conversion after
+        # get filepaths to process
+        if args.extension == ".parquet":
+            if args.from_tgz:
+                pass
+            for filepath_curr in sorted(glob.glob(os.path.join(folder_device_raw_current, "*.tif"))):
+                filename_curr = Path(filepath_curr).stem
+                if filename_curr not in made_parquets:
+                    filepaths_tif.append(filepath_curr)
+
+            # filepaths = filepaths[0:1]  # TODO delete not in debug
+
+            # first do parquet to csv, then tif - parquet - csv
+            if args.parquet_to_csv and not filepaths_parquet:
+                func = convert_parquet_to_csv
+                filepaths = filepaths_parquet
+                logger.debug(f"Process parquet_to_csv")
+            else:
+                func = convert_geotiff_to_parquet
+                filepaths = filepaths_tif
+                logger.debug(f"Process geotiff_to_parquet")
+        # only existing parquets to csv
+        elif args.extension == ".csv":
+            # + [Path(f).stem
+            #   for f in
+            #   list(glob.glob(os.path.join(folder_local_data, args.type, "result", "*" + ".csv")))]
+
+            for filepath_curr in sorted(glob.glob(os.path.join(folder_result, "*.parquet"))):
+                filename_curr = Path(filepath_curr).stem
+                if filename_curr not in made_csvs:
+                    filepaths_parquet.append(filepath_curr)
+            func = convert_parquet_to_csv
+            filepaths = filepaths_parquet
+            logger.debug(f"Process parquet_to_csv")
+        if not filepaths:
+            break
+        logger.debug(" ".join(filepaths))
+        # dates = ([re.findall(r'\d+', Path(f).stem)[0] for f in files])
         with parallel_executor as executor:
 
             future_to_url = {executor.submit(func, filepath, args):
@@ -272,14 +295,15 @@ def main(args):
                 try:
                     date = future.result()
                     dates.append(date)
-                    print(date)
+                    logger.info(date)
                 except Exception as exc:
                     print('%r generated an exception: %s' % (f, exc))
                 else:
                     print(f'{f} processed with result {date}')
 
         logger.info(f"Done dates {dates}")
-        logger.info(f"from {dates[0]} to {dates[-1]}")
+        if len(dates):
+            logger.info(f"from {dates[0]} to {dates[-1]}")
         # time.sleep(30 * 60)
 
 
@@ -295,15 +319,23 @@ if __name__ == '__main__':
 
     # remote device - existing folders
     folder_device_raw = r"Z:\Projekte\COVID19_Remote\01_raw_data"
+    # set general type, to find folder with raw .tif data on the remote device
+    if args.type in ("cloud_cover", "rade9d"):
+        t = "nightly"
+    else:
+        t = "monthly"
     # remote device - possibly new folders
+    folder_device_raw_current = os.path.join(folder_device_raw, t, args.type)
     folder_device_processed = r"Z:\Projekte\COVID19_Remote\02_processed_data\raster"
     folder_device_processed_parquet = os.path.join(folder_device_processed, "parquet")
     folder_device_processed_csv = os.path.join(folder_device_processed, "csv")
     folder_device_processed_parquet_current = os.path.join(folder_device_processed_parquet, args.type)
     folder_device_processed_csv_current = os.path.join(folder_device_processed_csv, args.type)
     for folder in (folder_temp, folder_logs, folder_result,
-                   folder_device_processed, folder_device_processed_csv, folder_device_processed_parquet,
-                   folder_device_processed_parquet_current, folder_device_processed_csv):
+                   folder_device_processed,
+                   folder_device_processed_csv, folder_device_processed_parquet,
+                   folder_device_processed_parquet_current, folder_device_processed_csv_current,
+                   folder_device_raw_current):
         os.makedirs(folder, exist_ok=True)
 
     global_log_file = os.path.join(folder_logs, args.type + time.strftime("_%Y%m%d_%H%M%S") + ".log")
@@ -317,7 +349,7 @@ if __name__ == '__main__':
     # xyz = gdal.Translate("dem2.xyz", ds)
     x1_0 = 40272
     x2_0 = 49921
-    y1_0= 2808
+    y1_0 = 2808
     y2_0 = 10441
     if args.type.startswith("vcm"):  # monthly
         # -60.0020833333500008,0.0020827333499938 : 59.9979176266500076,75.0020833333500008
@@ -327,7 +359,7 @@ if __name__ == '__main__':
         # account for this shift of the presented in .tif area:
         x1 = x1_0 - 240 * 120  # corner of zero: -180 -> -60
         x2 = x2_0 - 240 * 120
-        y1 = y1_0 # corner of zero: 75 -> 75
+        y1 = y1_0  # corner of zero: 75 -> 75
         y2 = y2_0
     else:  # the whole world for cloud_cover, rade9d
         # -180.0020833333499866, -65.0020844533500082: 180.0020862133499975, 75.0020833333500008
